Make `--dry-run` test `--ci-upload`.

+++ cmd/brew-test-bot.rb
@@ -970,7 +970,9 @@ module Homebrew
     bintray_user = ENV["HOMEBREW_BINTRAY_USER"]
     bintray_key = ENV["HOMEBREW_BINTRAY_KEY"]
     if !bintray_user || !bintray_key
-      raise "Missing HOMEBREW_BINTRAY_USER or HOMEBREW_BINTRAY_KEY variables!"
+      unless ARGV.include?("--dry-run")
+        raise "Missing HOMEBREW_BINTRAY_USER or HOMEBREW_BINTRAY_KEY variables!"
+      end
     end
 
     # Don't pass keys/cookies to subprocesses
@@ -983,11 +985,18 @@ module Homebrew
       jenkins = ENV["JENKINS_HOME"]
       job = ENV["UPSTREAM_JOB_NAME"]
       id = ENV["UPSTREAM_BUILD_ID"]
-      return unless jenkins
-      raise "Missing Jenkins variables!" if !job || !id
+      if !job || !id
+        unless ARGV.include?("--dry-run")
+          raise "Missing Jenkins variables!"
+        end
+      end
 
       bottles = Dir["#{jenkins}/jobs/#{job}/configurations/axis-version/*/builds/#{id}/archive/*.bottle*.*"]
-      return if bottles.empty?
+      if bottles.empty?
+        unless ARGV.include?("--dry-run")
+          raise "No bottles found!"
+        end
+      end
 
       FileUtils.cp bottles, Dir.pwd, verbose: true
     end
@@ -995,6 +1004,28 @@ module Homebrew
     json_files = Dir.glob("*.bottle.json")
     bottles_hash = json_files.reduce({}) do |hash, json_file|
       deep_merge_hashes hash, JSON.parse(IO.read(json_file))
+    end
+
+    if ARGV.include?("--dry-run")
+      bottles_hash = {
+        "testbottest" => {
+          "formula" => {
+            "pkg_version" => "1.0.0",
+          },
+          "bottle" => {
+            "tags" => {
+              Utils::Bottles.tag => {
+                "filename" => "testbottest-1.0.0.#{Utils::Bottles.tag}.bottle.tar.gz",
+                "sha256" => "20cdde424f5fe6d4fdb6a24cff41d2f7aefcd1ef2f98d46f6c074c36a1eef81e",
+              }
+            }
+          },
+          "bintray" => {
+            "package" => "testbottest",
+            "repository" => "bottles",
+          },
+        }
+      }
     end
 
     first_formula_name = bottles_hash.keys.first
@@ -1005,11 +1036,21 @@ module Homebrew
     ENV["GIT_WORK_TREE"] = tap.path
     ENV["GIT_DIR"] = "#{ENV["GIT_WORK_TREE"]}/.git"
 
-    quiet_system "git", "am", "--abort"
-    quiet_system "git", "rebase", "--abort"
-    safe_system "git", "checkout", "-f", "master"
-    safe_system "git", "reset", "--hard", "origin/master"
-    safe_system "brew", "update"
+    if ARGV.include?("--dry-run")
+      puts <<-EOS.undent
+        git am --abort
+        git rebase --abort
+        git checkout -f master
+        git reset --hard origin/master
+        brew update
+      EOS
+    else
+      quiet_system "git", "am", "--abort"
+      quiet_system "git", "rebase", "--abort"
+      safe_system "git", "checkout", "-f", "master"
+      safe_system "git", "reset", "--hard", "origin/master"
+      safe_system "brew", "update"
+    end
 
     # These variables are for Jenkins, Jenkins pipeline and
     # Circle CI respectively.
@@ -1021,8 +1062,10 @@ module Homebrew
 
     if ENV["UPSTREAM_BOTTLE_KEEP_OLD"] || ENV["BOT_PARAMS"].to_s.include?("--keep-old") || ARGV.include?("--keep-old")
       system "brew", "bottle", "--merge", "--write", "--keep-old", *json_files
-    else
+    elsif !ARGV.include?("--dry-run")
       system "brew", "bottle", "--merge", "--write", *json_files
+    else
+      puts "brew bottle --merge --write $JSON_FILES"
     end
 
     # These variables are for Jenkins and Circle CI respectively.
@@ -1034,9 +1077,16 @@ module Homebrew
       "testing-#{upstream_number}"
     elsif (number = ENV["BUILD_NUMBER"])
       "other-#{number}"
+    elsif ARGV.include?("--dry-run")
+      "$GIT_TAG"
     end
+
     if git_tag
-      safe_system "git", "push", "--force", remote, "master:master", ":refs/tags/#{git_tag}"
+      if ARGV.include?("--dry-run")
+        puts "git push --force #{remote} master:master :refs/tags/#{git_tag}"
+      else
+        safe_system "git", "push", "--force", remote, "master:master", ":refs/tags/#{git_tag}"
+      end
     end
 
     formula_packaged = {}
@@ -1051,12 +1101,17 @@ module Homebrew
       bottle_hash["bottle"]["tags"].each do |_tag, tag_hash|
         filename = tag_hash["filename"]
         bintray_filename_url = "https://api.bintray.com/file_version/#{bintray_org}/#{bintray_repo}/#{filename}"
-        filename_already_published = begin
-          output, = curl_output bintray_filename_url
-          json = JSON.parse output
-          json["published"]
-        rescue JSON::ParserError
+        filename_already_published = if ARGV.include?("--dry-run")
+          puts "curl #{bintray_filename_url}"
           false
+        else
+          begin
+            output, = curl_output bintray_filename_url
+            json = JSON.parse output
+            json["published"]
+          rescue JSON::ParserError
+            false
+          end
         end
 
         if filename_already_published
@@ -1068,31 +1123,61 @@ module Homebrew
 
         unless formula_packaged[formula_name]
           package_url = "#{bintray_packages_url}/#{bintray_package}"
-          unless system(*curl_args(extra_args: ["--output", "/dev/null", package_url]))
+          package_exists = if ARGV.include?("--dry-run")
+            puts "curl --output /dev/null #{package_url}"
+            false
+          else
+            system(*curl_args(extra_args: ["--output", "/dev/null", package_url]))
+          end
+
+          unless package_exists
             package_blob = <<-EOS.undent
               {"name": "#{bintray_package}",
                "public_download_numbers": true,
                "public_stats": true}
             EOS
-            curl "--user", "#{bintray_user}:#{bintray_key}",
-                 "--header", "Content-Type: application/json",
-                 "--data", package_blob, bintray_packages_url
-            puts
+            if ARGV.include?("--dry-run")
+              puts <<-EOS.undent
+                curl --user $BINTRAY_USER:$BINTRAY_KEY
+                     --header Content-Type: application/json
+                     --data #{package_blob.delete("\n")}
+                     #{bintray_packages_url}
+              EOS
+            else
+              curl "--user", "#{bintray_user}:#{bintray_key}",
+                   "--header", "Content-Type: application/json",
+                   "--data", package_blob, bintray_packages_url
+              puts
+            end
           end
           formula_packaged[formula_name] = true
         end
 
         content_url = "https://api.bintray.com/content/#{bintray_org}"
         content_url += "/#{bintray_repo}/#{bintray_package}/#{version}/#{filename}"
-        curl "--user", "#{bintray_user}:#{bintray_key}",
-             "--upload-file", filename, content_url
-        puts
+        if ARGV.include?("--dry-run")
+          puts <<-EOS.undent
+            curl --user $BINTRAY_USER:$BINTRAY_KEY
+                 --upload-file #{filename}
+                 #{content_url}
+          EOS
+        else
+          curl "--user", "#{bintray_user}:#{bintray_key}",
+               "--upload-file", filename, content_url
+          puts
+        end
       end
     end
 
     return unless git_tag
-    safe_system "git", "tag", "--force", git_tag
-    safe_system "git", "push", "--force", remote, "master:master", "refs/tags/#{git_tag}"
+
+    if ARGV.include?("--dry-run")
+      puts "git tag --force #{git_tag}"
+      puts "git push --force #{remote} master:master refs/tags/#{git_tag}"
+    else
+      safe_system "git", "tag", "--force", git_tag
+      safe_system "git", "push", "--force", remote, "master:master", "refs/tags/#{git_tag}"
+    end
   end
 
   def sanitize_argv_and_env
